requirements.txt
# Core dependencies for query engine benchmark
pyiceberg[hive]>=0.5.0
duckdb>=0.8.0
getdaft>=0.2.0
pymysql>=1.0.0
pyarrow>=12.0.0
psutil>=5.9.0
polars>=0.20.0

# Configuration management
pyyaml>=6.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
pytest-xdist>=3.0.0
ruff>=0.1.0
mypy>=1.0.0
